converted all message values to the binary format indicated in the spec

+++ networking/p2p/src/test/java/tech/pegasys/artemis/networking/p2p/hobbits/RPCCodecTest.java
@@ -43,8 +43,8 @@ final class RPCCodecTest {
   void testHello() {
     HelloMessage hello =
         new HelloMessage(
-            1,
-            1,
+            (short) 1,
+            (short) 1,
             Bytes32.random().toArrayUnsafe(),
             BigInteger.ZERO,
             Bytes32.random().toArrayUnsafe(),
